+++ app/controllers/api/v1/api_controller.rb
@@ -16,6 +16,11 @@ module Api::V1
 
     respond_to :json
 
+    # Doorkeeper methods
+    def current_resource_owner
+      User.find(doorkeeper_token.resource_owner_id) if doorkeeper_token
+    end
+
     protected
 
     # Devise methods
@@ -34,13 +39,6 @@ module Api::V1
     # Return the service version as a string
     def service_version
       '0.2.0'
-    end
-
-    private
-
-    # Doorkeeper methods
-    def current_resource_owner
-      User.find(doorkeeper_token.resource_owner_id) if doorkeeper_token
     end
   end
 end