+++ config/routes.rb
@@ -1,11 +1,8 @@
 Rails.application.routes.draw do
-  # Use for login and autorize all resource
-  use_doorkeeper do
-    # No need to register client application
-    skip_controllers :applications, :authorized_applications
-  end
-
-  root to: "home#index"
+  # Authentication and authorization resources
+  use_doorkeeper
+  devise_for :users
+  devise_for :admins
 
   # API documentation
   mount Rswag::Ui::Engine => '/api-docs'
@@ -13,18 +10,18 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
-      devise_for :users, controllers: {
-        registrations: 'api/v1/users/registrations',
-      }, skip: [:sessions, :password]
-      devise_for :admins, controllers: {
-        registrations: 'api/v1/admins/registrations',
-      }, skip: [:sessions, :password]
+      resources :profiles
+      resources :users
+      resources :admins
 
-      # API Active Check
-      get '/ping', to: 'ping#index'
-      # API Health Check
+      get "/me",   to: 'credentials#me'
+      get "/fast", to: 'fast#index'
       get '/status', to: 'status#index'
+      get '/ping', to: 'ping#index'
     end
-
   end
+
+  # Default landing page for service provider
+  get '/home', to: 'home#index'
+  root to: "home#index"
 end